+++ cpg_workflows/workflow.py
@@ -12,6 +12,7 @@ Examples of workflows can be found in the `production-workflows` repository.
 """
 
 import functools
+import hashlib
 import logging
 import pathlib
 from abc import ABC, abstractmethod
@@ -45,6 +46,43 @@ StageDecorator = Callable[..., 'Stage']
 # it would violate the Liskov substitution principle (i.e. any Stage subclass would
 # have to be able to work on any Target subclass).
 TargetT = TypeVar('TargetT', bound=Target)
+
+
+@functools.lru_cache(1)
+def alignment_inputs_hash_mc() -> str:
+    """
+    Unique hash string of sample alignment inputs. Useful to decide
+    whether the analysis on the target needs to be rerun.
+    """
+    all_sgs = get_multicohort().get_sequencing_groups()
+    s = ' '.join(sorted(' '.join(str(s.alignment_input)) for s in all_sgs if s.alignment_input))
+    return f'{hashlib.sha256(s.encode()).hexdigest()[:38]}_{len(all_sgs)}'
+
+
+@functools.lru_cache
+def cohort_inputs_hash_by_id(cohort_id: str) -> str:
+    """
+    Unique hash string of sample alignment inputs. Useful to decide
+    whether the analysis on the target needs to be rerun.
+    """
+    if (cohort := get_multicohort().get_cohort_by_name(cohort_id)) is None:
+        raise ValueError(f'Cohort {cohort_id} not found')
+    all_sgs = cohort.get_sequencing_groups()
+    s = ' '.join(sorted(' '.join(str(s.alignment_input)) for s in all_sgs if s.alignment_input))
+    return f'{hashlib.sha256(s.encode()).hexdigest()[:38]}_{len(all_sgs)}'
+
+
+@functools.lru_cache
+def dataset_inputs_hash_by_id(dataset_id: str) -> str:
+    """
+    Unique hash string of sample alignment inputs. Useful to decide
+    whether the analysis on the target needs to be rerun.
+    """
+    if (dataset := get_multicohort().get_dataset_by_name(dataset_id)) is None:
+        raise ValueError(f'Dataset {dataset_id} not found')
+    all_sgs = dataset.get_sequencing_groups()
+    s = ' '.join(sorted(' '.join(str(s.alignment_input)) for s in all_sgs if s.alignment_input))
+    return f'{hashlib.sha256(s.encode()).hexdigest()[:38]}_{len(all_sgs)}'
 
 
 def path_walk(expected, collected: set | None = None) -> set[Path]:
